requirements.txt

streamlit>=1.35.0
joblib>=1.3.2
numpy>=1.26.4
pandas>=2.1.4
matplotlib>=3.8.0
Pillow>=10.1.0


shap>=0.50.0
lightgbm>=4.6.0


charset-normalizer>=3.3.2